feat: Fetch historical weather for trips that are more than 1 year away
This feat will allow the app to returning the trip's year to the current
year, so it'll fetch the historical weather data of the last 3 years
prior to the current year. This is to prevent the historical weather API
to rejecting the request, since it only allows requests for dates prior
to the current date. E.g., if the trip was more than one year away, and
the app tried to fetch historical data of the last 3 years, it would end
up requesting future weather forecast, which is not allowed when using
historic weather API.